Removed lib doxygen abstract into separate files
This is to remove the need to ever include such documentation in a
header file. Placing the same (or worse, similar [or even worse --
outdated!!]) documentation in more than one header file results in nasty
Doxygen output. This way each lib's doxygen information is separate from
the code involved in the library.