libheartbeat is now modal on #define switch
with env option
libheartbeat will require linking with libos in order to provide each
thread a modal context with which to manipulate a muscle.

Without this switch no os integration is required and only control of
the colloquial heart is possible.

libheartbeat is contained entirely in the header file.

email=true