bugfixes: typos
tl;dr: crazy ranting that has nothing to do with you
due to persistent artifacts in the session used to create this stuff!
ugh i need more than just a rename-defun defun, i need rename-var and
rename-macro as well. or a global option to cover all the bases.. that
sounds sensible yeah