test-debounce, isr/thread synchronization works
When we though we were finished with our hardware driver library, we
were not. The function pointers in the will drivers always run in
priviledged mode. We must use mailboxes or FIFOs of some sort to be able
to communicate between user threads and kernel drivers, and we can use
semaphores for synchronization. We had the subscription idea down, but
we were missing both the mailbox and semaphore constructs. We
fundamentally were not finished with libhw.

The button debouncer is about to become its own driver which utilizes
the subscription capabilities to listen to interesting GPIO ports. The
function pointers it passes in act as merely data sanitizers to clean up
the data and put it in a more presentable form for the running user
threads, e.g. the "CLIENT CODE" in postpone_suicide() function.

email=true